test: isolate tall preview fit regression
